// planning-idsdp-pps-frontend/src/app/main/modules/dpp-tapp-management/models/logFrame.model.ts
export class LogFrameModel {
    id: string;
    uuid: string;
    projectConceptMasterId: number;
    projectConceptUuid: string;
    dppMasterId: number;
    plannedDate: string;
    dateOfLogFrame: string;
    goalNS: string;
    goalOVI: string;
    goalMOV: boolean;
    objectiveNS: boolean;
    objectiveOVI: boolean;
    objectiveMOV: boolean;
    objectiveIA: boolean;
    outputNS: boolean;
    outputOVI: boolean;
    outputMOV: boolean;
    outputIA: boolean;
    inputNS: boolean;
    inputOVI: boolean;
    inputMOV: boolean;
    inputIA: boolean;

}